+++ server/src/main/java/org/opensearch/common/settings/IndexScopedSettings.java
@@ -283,6 +283,10 @@ public final class IndexScopedSettings extends AbstractScopedSettings {
                 // Settings for Auto Force Merge
                 IndexSettings.INDEX_AUTO_FORCE_MERGES_ENABLED,
 
+                // Setting for derived source feature
+                IndexSettings.INDEX_DERIVED_SOURCE_SETTING,
+                IndexSettings.INDEX_DERIVED_SOURCE_TRANSLOG_ENABLED_SETTING,
+
                 // validate that built-in similarities don't get redefined
                 Setting.groupSetting("index.similarity.", (s) -> {
                     Map<String, Settings> groups = s.getAsGroups();